+++ 012/ui/blank_control.go
@@ -9,12 +9,12 @@ type BlankControl struct {
 	parts.ControlBase
 }
 
-func NewBlankControl(x, y, w, h int) *BlankControl {
+func NewBlankControl(x, y, w, h float64) *BlankControl {
 	b := &BlankControl{}
 	b.InitBlankControl(x, y, w, h)
 	return b
 }
 
-func (b *BlankControl) InitBlankControl(x, y, w, h int) {
+func (b *BlankControl) InitBlankControl(x, y, w, h float64) {
 	b.InitControlBase(b, x, y, w, h)
 }
