+++ 012/ui/grouping_control.go
@@ -8,13 +8,13 @@ type GroupingControl struct {
 	parts.Grouping
 }
 
-func NewGroupingControl(x, y, w, h int) *GroupingControl {
+func NewGroupingControl(x, y, w, h float64) *GroupingControl {
 	c := &GroupingControl{}
 	c.InitGroupingControl(x, y, w, h)
 	return c
 }
 
-func (c *GroupingControl) InitGroupingControl(x, y, w, h int) {
+func (c *GroupingControl) InitGroupingControl(x, y, w, h float64) {
 	c.InitControlBase(c, x, y, w, h)
 	c.InitGrouping(c)
 }
