+++ 012/ui/label.go
@@ -10,13 +10,13 @@ type Label struct {
 	parts.TextDrawable
 }
 
-func NewLabel(x, y, w, h int, text string, size int) *Label {
+func NewLabel(x, y, w, h float64, text string, size float64) *Label {
 	l := &Label{}
 	l.InitLabel(x, y, w, h, text, size)
 	return l
 }
 
-func (l *Label) InitLabel(x, y, w, h int, text string, size int) {
+func (l *Label) InitLabel(x, y, w, h float64, text string, size float64) {
 	l.InitControlBase(l, x, y, w, h)
 	l.InitTextDrawable(l, text, size, parts.AlignLeft, parts.AlignCenter, 0, 0, color.White, true)
 }
