+++ 012/ui/window.go
@@ -13,27 +13,27 @@ type TitleBar struct {
 
 // NewTitleBarは新しいタイトルバーを作成する
 // ドラッグ操作によるウィンドウ移動のイベントハンドラを設定する
-func NewTitleBar(x, y, h int, text string) *TitleBar {
+func NewTitleBar(x, y, h float64, text string) *TitleBar {
 	t := &TitleBar{}
 	t.InitTitleBar(x, y, h, text)
 	return t
 }
 
-func (t *TitleBar) InitTitleBar(x, y, h int, text string) {
+func (t *TitleBar) InitTitleBar(x, y, h float64, text string) {
 	theme := parts.CurrentTheme
 	t.InitInteractiveControl(x, y, 0, h)
 	t.InitTextDrawable(t, text, h*2/3, parts.AlignCenter, parts.AlignCenter, 0, 0, theme.Text, true)
 	t.BackColor = theme.TitleBar
 
 	// TitleBarをドラッグすると親のWindowが移動するように設定
-	var dragOffsetX, dragOffsetY int
-	t.OnDragStart = func(x, y int) {
+	var dragOffsetX, dragOffsetY float64
+	t.OnDragStart = func(x, y float64) {
 		// 親のWindowとの相対座標を保存
 		cb := t.Parent.GetEntityBase()
 		dragOffsetX = x - cb.X
 		dragOffsetY = y - cb.Y
 	}
-	t.OnDrag = func(x, y int) {
+	t.OnDrag = func(x, y float64) {
 		// 親のWindowの座標を更新
 		cb := t.Parent.GetEntityBase()
 		cb.X = x - dragOffsetX
@@ -50,27 +50,27 @@ type ClientArea struct {
 
 // NewClientAreaは新しいクライアント領域を作成する
 // ウィンドウのドラッグ移動に対応する
-func NewClientArea(x, y int) *ClientArea {
+func NewClientArea(x, y float64) *ClientArea {
 	c := &ClientArea{}
 	c.InitClientArea(x, y)
 	return c
 }
 
-func (c *ClientArea) InitClientArea(x, y int) {
+func (c *ClientArea) InitClientArea(x, y float64) {
 	c.InitInteractiveControl(x, y, 0, 0)
 	c.InitGrouping(c)
 	c.BackColor = parts.CurrentTheme.ClientArea
 	c.FlexGrow = 1
 
 	// ClientAreaをドラッグすると親のWindowが移動する
-	var dragOffsetX, dragOffsetY int
-	c.OnDragStart = func(x, y int) {
+	var dragOffsetX, dragOffsetY float64
+	c.OnDragStart = func(x, y float64) {
 		// 親のWindowとの相対座標を保存
 		cb := c.Parent.GetEntityBase()
 		dragOffsetX = x - cb.X
 		dragOffsetY = y - cb.Y
 	}
-	c.OnDrag = func(x, y int) {
+	c.OnDrag = func(x, y float64) {
 		// 親のWindowの座標を更新
 		cb := c.Parent.GetEntityBase()
 		cb.X = x - dragOffsetX
@@ -91,7 +91,7 @@ type Window struct {
 
 // NewWindowは新しいウィンドウを作成する
 // タイトルバーとクライアント領域を初期化し、レイアウトを設定する
-func NewWindow(x, y, w, h int, text string) *Window {
+func NewWindow(x, y, w, h float64, text string) *Window {
 	win := &Window{}
 	win.InitControlBase(win, x, y, w, h)
 	win.InitResizable(win)
